Move middleware to transport/http
